+++ CHANGELOG.md
@@ -5,14 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- List games page
+- Add how to play instructions for each game
+
+### Changed
+
+- Replace all `multi-hangman` with `super-hangman`
+
 ## [0.2.0] - 2024-08-16
 
 ### Added
 
-Games
-
 - Add super tic-tac-toe game
-- Rename multi hangman to super hangman
+
+### Changed
+
+- Name the Multi Hangman game as Super Hangman
 
 ## [0.1.1] - 2024-08-11